Чистка кода
