+++ src/main/java/ru/biderman/s1000dpmviewer/S1000dPmViewerApplication.java
@@ -11,8 +11,3 @@ public class S1000dPmViewerApplication {
     }
 
 }
-
-// TODO дополнительное поле: кто добавил публикацию
-
-// TODO разобраться, нельзя ли сделать в Publication поле типа XML (Document)
-// TODO попробовать отделить интеграционные тесты на postgres?
