--- src/main/java/ru/biderman/s1000dpmviewer/exceptions/ServerErrorException.java
@@ -1,7 +1,0 @@
-package ru.biderman.s1000dpmviewer.exceptions;
-
-public class ServerErrorException extends Exception {
-    public ServerErrorException(Throwable cause) {
-        super(cause);
-    }
-}
