+++ src/main/java/ru/biderman/s1000dpmviewer/xmlparsers/PublicationParser.java
@@ -2,7 +2,6 @@ package ru.biderman.s1000dpmviewer.xmlparsers;
 
 import org.w3c.dom.Document;
 import ru.biderman.s1000dpmviewer.domain.Publication;
-import ru.biderman.s1000dpmviewer.domain.PublicationDetails;
 import ru.biderman.s1000dpmviewer.domain.publicationcontent.Entry;
 import ru.biderman.s1000dpmviewer.exceptions.InvalidPublicationException;
 
@@ -10,6 +9,5 @@ import java.io.InputStream;
 
 public interface PublicationParser {
     Publication createPublication(InputStream inputStream) throws InvalidPublicationException;
-    PublicationDetails getPublicationDetails(Document document) throws InvalidPublicationException;
     Entry getPublicationContent(Document document);
 }
