+++ src/ui/src/components/publication/PublicationAdd.vue
@@ -43,8 +43,8 @@
             loading: false,
             loadingErrorText: null,
             fileRules: [
-                v => v || 'Не выбран файл для загрузки.',
-                v => v && v.size < 1048576 || 'Файл с публикацией не должен превышать 1 Мб.'
+                v => !!v || 'Не выбран файл для загрузки.',
+                v => !!v && v.size < 1048576 || 'Файл с публикацией не должен превышать 1 Мб.'
             ]
         }),
